requirements.txt
allure-python-commons==2.14.3
annotated-types==0.7.0
attrs==25.3.0
bidict==0.23.1
blinker==1.9.0
Brotli==1.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
ConfigArgParse==1.7.1
Faker==37.4.2
Flask==3.1.1
flask-cors==6.0.1
Flask-Login==0.6.3
gevent==25.5.1
geventhttpclient==2.3.4
greenlet==3.2.3
h11==0.16.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
locust==2.37.14
locust-cloud==1.26.3
MarkupSafe==3.0.2
msgpack==1.1.1
outcome==1.3.0.post0
packaging==25.0
platformdirs==4.3.8
pluggy==1.6.0
psutil==7.0.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PySocks==1.7.1
pytest==8.4.1
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
PyYAML==6.0.2
pyzmq==27.0.0
requests==2.32.4
selenium==4.33.0
setuptools==80.9.0
simple-websocket==1.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
zope.event==5.1
zope.interface==7.2
